Align verification contracts with SDK conventions
Keep DTOs and enums in dedicated files and use standard spaced enum values for stable extension.
