+++ TLabs.ExchangeSdk/Verification/VerificationDataSource.cs
@@ -1,0 +1,8 @@
+namespace TLabs.ExchangeSdk.Verification;
+
+public enum VerificationDataSource
+{
+    OldFrontend = 10,
+    Admin = 20,
+    Sumsub = 30,
+}
